Change `PropertyIndexer.m_propDict` key from int to `eProperty` Replaced usages of the int indexer. Marked it obsolete. Changed `GamePlayer.RefreshItemBonuses` to clear and reuse the existing collection.

+++ GameServer/ECS-Effects/DirtyTricksECSEffect.cs
@@ -91,7 +91,7 @@ namespace DOL.GS
 
         public override void OnStartEffect()
         {
-            Owner.DebuffCategory[(int)eProperty.FumbleChance] += 35;
+            Owner.DebuffCategory[eProperty.FumbleChance] += 35;
 
             if (OwnerPlayer != null)
             {
@@ -106,7 +106,7 @@ namespace DOL.GS
         }
         public override void OnStopEffect()
         {
-            Owner.DebuffCategory[(int)eProperty.FumbleChance] -= 35;
+            Owner.DebuffCategory[eProperty.FumbleChance] -= 35;
 
             if (OwnerPlayer != null)
             {
