+++ GameServer/propertycalc/ArcherySpeedPercentCalculator.cs
@@ -35,13 +35,13 @@ namespace DOL.GS.PropertyCalc
 		public override int CalcValue(GameLiving living, eProperty property)
 		{
 			int archerySpeed = 0;
-			archerySpeed +=  living.BaseBuffBonusCategory[(int)property] 
-			                 + living.SpecBuffBonusCategory[(int)property] 
-			                 - living.DebuffCategory[(int)property] 
-			                 + living.OtherBonus[(int)property] 
-			                 + living.AbilityBonus[(int)property] ;
+			archerySpeed +=  living.BaseBuffBonusCategory[property] 
+			                 + living.SpecBuffBonusCategory[property] 
+			                 - living.DebuffCategory[property] 
+			                 + living.OtherBonus[property] 
+			                 + living.AbilityBonus[property] ;
 			//hardcap at 10%
-			//return Math.Min(10, living.ItemBonus[(int)property] - living.DebuffCategory[(int)property]);
+			//return Math.Min(10, living.ItemBonus[property] - living.DebuffCategory[property]);
 
 			return archerySpeed;
 		}
