+++ GameServer/propertycalc/HealthRegenerationAmountCalculator.cs
@@ -32,12 +32,12 @@ namespace DOL.GS.PropertyCalc
             // From DoL's `living.Level * 0.6` above level 25, `10 + (living.Level * 0.2)` below level 26.
             // 15 health per tick at level 50 instead of 30.
             double regen = 2.5 + living.Level * 0.25;
-            int debuff = living.SpecBuffBonusCategory[(int) property];
+            int debuff = living.SpecBuffBonusCategory[property];
 
             if (debuff < 0)
                 debuff = -debuff;
 
-            regen += living.BaseBuffBonusCategory[(int) property] + living.AbilityBonus[(int) property] + living.ItemBonus[(int)property] - debuff;
+            regen += living.BaseBuffBonusCategory[property] + living.AbilityBonus[property] + living.ItemBonus[property] - debuff;
 
             if (living is GameNPC npc)
             {
