+++ GameServer/propertycalc/WeaponSkillCalculator.cs
@@ -7,11 +7,11 @@ namespace DOL.GS.PropertyCalc
 		public override int CalcValue(GameLiving living, eProperty property)
 		{
 			double percent = 100
-			+ living.BaseBuffBonusCategory[(int)property] // enchance the weaponskill
-			+ living.SpecBuffBonusCategory[(int)property] // enchance the weaponskill
+			+ living.BaseBuffBonusCategory[property] // enchance the weaponskill
+			+ living.SpecBuffBonusCategory[property] // enchance the weaponskill
 				//hotfix for poisons where both debuff components have same value
-			- (int)Math.Abs(living.DebuffCategory[(int)property] / 5.4) // reduce
-			+ living.ItemBonus[(int)property];
+			- (int)Math.Abs(living.DebuffCategory[property] / 5.4) // reduce
+			+ living.ItemBonus[property];
 			return (int)Math.Max(1, percent);
 		}
 	}
